# workflow/openapi.yaml
...
openapi: 3.1.0
info:
  title: GW/AP Debug Platform Workflow API
  version: 0.9.0
servers:
- url: http://localhost:8000/api/v1
...
paths:
...
  /knowledge/upload:
    post:
      summary: Upload Knowledge
      operationId: uploadKnowledgeDocument
      parameters:
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_knowledge_api_v1_knowledge_upload_post'
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/KnowledgeImportOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      x-agent-callable: true
      description: ADMIN/EXPERT direct knowledge or Skill import. ENGINEER ordinary Markdown uploads use knowledge-contributions/upload
        and cannot turn an ordinary upload into a mandatory Skill.
...
  /workbench/knowledge:
    get:
      tags:
      - workbench
      summary: Knowledge
      operationId: knowledge_api_v1_workbench_knowledge_get
      parameters:
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      - VIEWER
      x-agent-callable: true
...
  /knowledge-contributions:
    get:
      tags:
      - knowledge-contributions
      summary: List Contributions
      operationId: list_contributions_api_v1_knowledge_contributions_get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: content_kind
        in: query
        required: false
        schema:
          anyOf:
          - enum:
            - KNOWLEDGE
            - SKILL
            type: string
          - type: 'null'
          title: Content Kind
      - name: mine
        in: query
        required: false
        schema:
          type: boolean
          default: false
          title: Mine
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 500
          minimum: 1
          default: 100
          title: Limit
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      x-agent-callable: false
      x-resource-scope: own_drafts_and_submissions_or_manager_review_queue
      description: Return caller-owned drafts/submissions, or an ADMIN/EXPERT review queue excluding other users
        private DRAFTs. No contribution is shared merely by being listed.
    post:
      tags:
      - knowledge-contributions
      summary: Create
      operationId: create_api_v1_knowledge_contributions_post
      parameters:
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContributionCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      x-agent-callable: false
      x-resource-scope: own_unpublished_proposal
      description: Create an owned unpublished ordinary-knowledge proposal, shared-content change/deletion request,
        or Skill amendment. Ownership does not grant direct Skill management or publication.
  /knowledge-contributions/upload:
    post:
      tags:
      - knowledge-contributions
      summary: Upload
      operationId: upload_api_v1_knowledge_contributions_upload_post
      parameters:
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_api_v1_knowledge_contributions_upload_post'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      x-agent-callable: false
      x-resource-scope: own_ordinary_knowledge_draft_even_if_filename_is_SKILL_md
      description: Upload UTF-8 Markdown into an owned ordinary-knowledge draft. A file named SKILL.md still has
        content_kind=KNOWLEDGE and does not become a mandatory diagnostic Skill.
  /knowledge-contributions/from-library/{record_id}:
    post:
      tags:
      - knowledge-contributions
      summary: From Library
      operationId: from_library_api_v1_knowledge_contributions_from_library__record_id__post
      parameters:
      - name: record_id
        in: path
        required: true
        schema:
          type: string
          title: Record Id
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      x-agent-callable: false
      description: Create an owned knowledge proposal from a visible eligible library record; any shared publication
        requires ADMIN/EXPERT review.
  /knowledge-contributions/{contribution_id}:
    get:
      tags:
      - knowledge-contributions
      summary: Get
      operationId: get_api_v1_knowledge_contributions__contribution_id__get
      parameters:
      - name: contribution_id
        in: path
        required: true
        schema:
          type: string
          title: Contribution Id
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      x-agent-callable: false
      x-resource-scope: owner_or_submitted_manager_review
      description: Read an owned proposal or an ADMIN/EXPERT-authorized submitted proposal. Other users private
        drafts are hidden.
    patch:
      tags:
      - knowledge-contributions
      summary: Edit
      operationId: edit_api_v1_knowledge_contributions__contribution_id__patch
      parameters:
      - name: contribution_id
        in: path
        required: true
        schema:
          type: string
          title: Contribution Id
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContributionUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      x-agent-callable: false
      x-resource-scope: authenticated_owner_and_editable_draft_state
      description: The owner edits an editable unpublished proposal with expected_version; shared documents and
        mandatory Skills are not directly changed.
    delete:
      tags:
      - knowledge-contributions
      summary: Delete
      operationId: delete_api_v1_knowledge_contributions__contribution_id__delete
      parameters:
      - name: contribution_id
        in: path
        required: true
        schema:
          type: string
          title: Contribution Id
      - name: expected_version
        in: query
        required: true
        schema:
          type: integer
          minimum: 1
          title: Expected Version
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      x-agent-callable: false
      x-resource-scope: authenticated_owner_and_expected_version
      description: The owner deletes an eligible proposal with expected_version. Shared-document deletion is a separate
        proposal requiring review.
  /knowledge-contributions/{contribution_id}/submit:
    post:
      tags:
      - knowledge-contributions
      summary: Submit
      operationId: submit_api_v1_knowledge_contributions__contribution_id__submit_post
      parameters:
      - name: contribution_id
        in: path
        required: true
        schema:
          type: string
          title: Contribution Id
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContributionVersion'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      - ENGINEER
      x-agent-callable: false
      x-resource-scope: authenticated_owner_and_expected_version
      description: The owner submits the current proposal version for ADMIN/EXPERT review. Submission does not itself
        publish shared knowledge.
  /knowledge-contributions/{contribution_id}/review-draft:
    patch:
      tags:
      - knowledge-contributions
      summary: Reviewer Edit
      operationId: reviewer_edit_api_v1_knowledge_contributions__contribution_id__review_draft_patch
      parameters:
      - name: contribution_id
        in: path
        required: true
        schema:
          type: string
          title: Contribution Id
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContributionUpdate'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      x-agent-callable: false
      description: ADMIN/EXPERT correct submitted proposal content using expected_version. Preserve the original
        submission, version differences, and review history.
  /knowledge-contributions/{contribution_id}/review-chat:
    post:
      tags:
      - knowledge-contributions
      summary: Reviewer Chat
      operationId: reviewer_chat_api_v1_knowledge_contributions__contribution_id__review_chat_post
      parameters:
      - name: contribution_id
        in: path
        required: true
        schema:
          type: string
          title: Contribution Id
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContributionChat'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      x-agent-callable: false
      description: ADMIN/EXPERT use multiple AI turns to correct a submitted proposal. Use the reviewer visible
        selected/personal/shared Chat profile; do not consume the submitter private model. Corrections remain unpublished
        until review.
  /knowledge-contributions/{contribution_id}/review:
    post:
      tags:
      - knowledge-contributions
      summary: Review
      operationId: review_api_v1_knowledge_contributions__contribution_id__review_post
      parameters:
      - name: contribution_id
        in: path
        required: true
        schema:
          type: string
          title: Contribution Id
      - name: x-api-key
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: X-Api-Key
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContributionReview'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '403':
          description: Role, ownership, or resource scope does not authorize this operation
      x-allowed-roles:
      - ADMIN
      - EXPERT
      x-agent-callable: false
      description: ADMIN/EXPERT approve or reject the exact current proposal version. Approval is publication approval;
        a successful atomic index generation makes the approved content shared, and failed publication preserves
        the previous active content.
components:
  schemas:
...
    Body_upload_api_v1_knowledge_contributions_upload_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        category_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Category Id
      type: object
      required:
      - file
      title: Body_upload_api_v1_knowledge_contributions_upload_post
...
    Body_upload_knowledge_api_v1_knowledge_upload_post:
      properties:
        file:
          type: string
          contentMediaType: application/octet-stream
          title: File
        source_type:
          type: string
          title: Source Type
          default: document
        device_type:
          anyOf:
          - type: string
          - type: 'null'
          title: Device Type
        module:
          anyOf:
          - type: string
          - type: 'null'
          title: Module
        trust_level:
          type: string
          title: Trust Level
          default: MEDIUM
        category_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Category Id
      type: object
      required:
      - file
      title: Body_upload_knowledge_api_v1_knowledge_upload_post
...
    ContributionChat:
      properties:
        expected_version:
          type: integer
          minimum: 1.0
          title: Expected Version
        instruction:
          type: string
          maxLength: 16000
          minLength: 1
          title: Instruction
        model_profile_id:
          anyOf:
          - type: string
            maxLength: 40
          - type: 'null'
          title: Model Profile Id
        consent_model_egress:
          type: boolean
          title: Consent Model Egress
          default: true
      additionalProperties: false
      type: object
      required:
      - expected_version
      - instruction
      title: ContributionChat
    ContributionCreate:
      properties:
        operation:
          type: string
          enum:
          - CREATE
          - UPDATE
          - DELETE
          title: Operation
          default: CREATE
        content_kind:
          type: string
          enum:
          - KNOWLEDGE
          - SKILL
          title: Content Kind
          default: KNOWLEDGE
        target_document_id:
          anyOf:
          - type: string
            maxLength: 40
          - type: 'null'
          title: Target Document Id
        title:
          anyOf:
          - type: string
            maxLength: 512
            minLength: 1
          - type: 'null'
          title: Title
        content:
          anyOf:
          - type: string
            maxLength: 500000
          - type: 'null'
          title: Content
        source_type:
          anyOf:
          - type: string
            maxLength: 64
          - type: 'null'
          title: Source Type
        category_id:
          anyOf:
          - type: string
            maxLength: 80
          - type: 'null'
          title: Category Id
        metadata:
          additionalProperties: true
          type: object
          title: Metadata
        source_curation_id:
          anyOf:
          - type: string
            maxLength: 40
          - type: 'null'
          title: Source Curation Id
      additionalProperties: false
      type: object
      title: ContributionCreate
    ContributionReview:
      properties:
        expected_version:
          type: integer
          minimum: 1.0
          title: Expected Version
        expected_content_hash:
          type: string
          pattern: ^[a-f0-9]{64}$
          title: Expected Content Hash
        action:
          type: string
          enum:
          - APPROVE
          - REJECT
          - RETURN
          title: Action
        comment:
          type: string
          maxLength: 4000
          title: Comment
          default: ''
      additionalProperties: false
      type: object
      required:
      - expected_version
      - expected_content_hash
      - action
      title: ContributionReview
    ContributionUpdate:
      properties:
        expected_version:
          type: integer
          minimum: 1.0
          title: Expected Version
        title:
          anyOf:
          - type: string
            maxLength: 512
            minLength: 1
          - type: 'null'
          title: Title
        content:
          anyOf:
          - type: string
            maxLength: 500000
          - type: 'null'
          title: Content
        category_id:
          anyOf:
          - type: string
            maxLength: 80
          - type: 'null'
          title: Category Id
        metadata:
          anyOf:
          - additionalProperties: true
            type: object
          - type: 'null'
          title: Metadata
        confidentiality:
          anyOf:
          - type: string
            enum:
            - PUBLIC
            - INTERNAL
            - RESTRICTED
          - type: 'null'
          title: Confidentiality
        comment:
          type: string
          maxLength: 4000
          title: Comment
          default: ''
      additionalProperties: false
      type: object
      required:
      - expected_version
      title: ContributionUpdate
    ContributionVersion:
      properties:
        expected_version:
          type: integer
          minimum: 1.0
          title: Expected Version
      additionalProperties: false
      type: object
      required:
      - expected_version
      title: ContributionVersion
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    JobOut:
      properties:
        id:
          type: string
          title: Id
        kind:
          type: string
          title: Kind
        status:
          type: string
          title: Status
        progress:
          type: integer
          title: Progress
        message:
          type: string
          title: Message
        result_json:
          type: string
          title: Result Json
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        idempotency_key:
          anyOf:
          - type: string
          - type: 'null'
          title: Idempotency Key
        attempt:
          type: integer
          title: Attempt
        max_attempts:
          type: integer
          title: Max Attempts
        available_at:
          type: string
          format: date-time
          title: Available At
        lease_owner:
          anyOf:
          - type: string
          - type: 'null'
          title: Lease Owner
        lease_expires_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Lease Expires At
        heartbeat_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Heartbeat At
        deadline_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Deadline At
        timeout_seconds:
          type: integer
          title: Timeout Seconds
        resource_limits_json:
          type: string
          title: Resource Limits Json
        dead_letter_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Dead Letter At
        dead_letter_reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Dead Letter Reason
        created_at:
          type: string
          format: date-time
          title: Created At
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
      type: object
      required:
      - id
      - kind
      - status
      - progress
      - message
      - result_json
      - error_message
      - idempotency_key
      - attempt
      - max_attempts
      - available_at
      - lease_owner
      - lease_expires_at
      - heartbeat_at
      - deadline_at
      - timeout_seconds
      - resource_limits_json
      - dead_letter_at
      - dead_letter_reason
      - created_at
      - started_at
      - completed_at
      title: JobOut
...
    KnowledgeImportOut:
      properties:
        document_id:
          type: string
          title: Document Id
        artifact_id:
          type: string
          title: Artifact Id
        job:
          $ref: '#/components/schemas/JobOut'
      type: object
      required:
      - document_id
      - artifact_id
      - job
      title: KnowledgeImportOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      name: X-API-Key
security:
- ApiKey: []